fix builder errors

+++ components/BottomAppBar/src/private/MDCBottomAppBarLayer.m
@@ -45,10 +45,10 @@
   CGFloat arcRadius =
       CGRectGetHeight(floatingButton.bounds) / 2 + kMDCBottomAppBarFloatingButtonRadiusOffset;
   CGFloat navigationBarYOffset = CGRectGetMinY(navigationBarFrame);
-  CGFloat halfAngle = acos((navigationBarYOffset - floatingButton.center.y) / arcRadius);
-  CGFloat startAngle = M_PI / 2 + halfAngle;
-  CGFloat endAngle = M_PI / 2 - halfAngle;
-  CGFloat halfOfHypotenuseLength = sin(halfAngle) * arcRadius;
+  CGFloat halfAngle = acosf((float)((navigationBarYOffset - floatingButton.center.y) / arcRadius));
+  CGFloat startAngle = M_PI / 2.0f + halfAngle;
+  CGFloat endAngle = M_PI / 2.0f - halfAngle;
+  CGFloat halfOfHypotenuseLength = sinf((float)halfAngle) * arcRadius;
 
   CGFloat width = CGRectGetWidth(rect);
   CGFloat height = CGRectGetHeight(rect);
